memccached: refactor to get one access per get instead of 4
- support dep files
- manifest files stored in fscache

performance is now ~ the same order as fsccache

run with bench.sh attached in the commit:
mkdir -p testbin
cp ccache testbin
make clean >/dev/null
rm -rf ~/.ccache
echo cold fsccache
time make CC="./testbin/ccache gcc" >/dev/null
make clean >/dev/null
echo hot fsccache
time make CC="./testbin/ccache gcc" >/dev/null
rm -rf ~/.ccache
export CCACHE_MEMCACHED_CONF=--SERVER=localhost:22122
memcached -p 22122 &
echo cold memccached
make clean >/dev/null
time make CC="./testbin/ccache gcc" >/dev/null
make clean >/dev/null
echo hot memccached
time make CC="./testbin/ccache gcc" >/dev/null
memstat --servers=localhost:22122
kill %1

cold fsccache
real	0m3.820s
user	0m2.956s
sys	0m0.544s

hot fsccache
real	0m0.204s
user	0m0.076s
sys	0m0.048s

cold memccached
real	0m3.813s
user	0m2.880s
sys	0m0.576s

hot memccached
real	0m0.285s
user	0m0.136s
sys	0m0.072s

+++ test.sh
@@ -57,7 +57,10 @@ test_failed() {
     echo TEST FAILED
     echo "Test data and log file have been left in $TESTDIR"
     tail -n 50 $CCACHE_LOGFILE
-    kill %1
+    if [ ! -z $CCACHE_MEMCACHED_CONF ]; then
+        memstat --servers=localhost:22122
+        kill %1
+    fi
     exit 1
 }
